working on login logout system of django@
cls

+++ base/views.py
@@ -1,9 +1,13 @@
 from django.shortcuts import render, redirect
-
 from .models import Room, User, Topic
-
+from django.contrib import messages
 from .forms import RoomForm
+from django.http import HttpResponse
 
+"""for authentication"""
+from django.contrib.auth import authenticate, login, logout
+from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 
 # rooms = [
 #     {"id": 1, "name": "lets learn python"},
@@ -33,6 +37,7 @@ def room(request, pk):
     return render(request, "base/room.html", context)
 
 
+@login_required(login_url="/login")
 def create_room(request):
     if request.method == "POST":
         print(request.POST)
@@ -45,9 +50,14 @@ def create_room(request):
     return render(request, "base/room_form.html", {"form": form})
 
 
+@login_required(login_url="/login")
 def update_room(request, pk):
+
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("you are not allowed here")
 
     context = {"form": form}
 
@@ -61,9 +71,62 @@ def update_room(request, pk):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="/login")
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("you are not allowed here")
     if request.method == "POST":
         room.delete()
         return redirect("home")
     return render(request, "base/delete_room.html")
+
+
+def login_page(request):
+    page = "login"
+    if request.method == "POST":
+        username = request.POST.get("username")
+        password = request.POST.get("password")
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.add_message(request, messages.ERROR, "User does not exist")
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.add_message(
+                request, messages.ERROR, "User name or pass is wrong exist"
+            )
+
+    context = {"page": page}
+    return render(request, "base/login_register.html", context)
+
+
+def logout_user(request):
+
+    logout(request)
+    return redirect("home")
+
+
+def register_user(request):
+    page = "register"
+    return render(request, "base/login_register.html")
+
+
+# def login_page(request):
+
+#     if request.method == "POST":
+#         username = request.POST.get("username")
+#         password = request.POST.get("password")
+#         try:
+#             user = User.objects.get(username=username)
+#         except:
+#             messages.add_messagee
+#             messages.add_message(request, messages.ERROR, "User does not exist")
+
+
+#     context = {}
+#     return render(request, "base/login_register.html", context)
